add Chatbox.js
